+++ miden/src/tools/mod.rs
@@ -34,32 +34,31 @@ impl Analyze {
         let stack_inputs = input_data.parse_stack_inputs()?;
         let advice_provider = input_data.parse_advice_provider()?;
 
-        let program_info: ProgramInfo = analyze(program.as_str(), stack_inputs, advice_provider)
-            .expect("Could not retrieve program info");
+        let execution_details: ExecutionDetails =
+            analyze(program.as_str(), stack_inputs, advice_provider)
+                .expect("Could not retrieve execution details");
 
-        println!("{}", program_info);
+        println!("{}", execution_details);
 
         Ok(())
     }
 }
 
-// PROGRAM INFO
+// EXECUTION DETAILS
 // ================================================================================================
 
-/// Contains info of a program. Used for program analysis. Contains the following fields:
-/// - total_vm_cycles: vm cycles it takes to execute the entire program
-/// - total_noops: total noops executed as part of a program
-/// - asm_op_stats: vector of [AsmOpStats] that contains assembly instructions and
-///   the number of vm cycles it takes to execute the instruction and the number of times the
-///   instruction is run as part of the given program.
+/// Contains details of executing a program, used for program analysis.
 #[derive(Debug, Default, Eq, PartialEq)]
-pub struct ProgramInfo {
+pub struct ExecutionDetails {
+    /// Number of VM cycles it took to execute the entire program.
     total_vm_cycles: u32,
+    /// Number of noops executed as part of a program.
     total_noops: usize,
+    /// Statistics about individual assembly operations executed by the VM, see [AsmOpStats].
     asm_op_stats: Vec<AsmOpStats>,
 }
 
-impl ProgramInfo {
+impl ExecutionDetails {
     /// Returns total vm cycles to execute a program
     pub fn total_vm_cycles(&self) -> u32 {
         self.total_vm_cycles
@@ -119,7 +118,7 @@ impl ProgramInfo {
     }
 }
 
-impl fmt::Display for ProgramInfo {
+impl fmt::Display for ExecutionDetails {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let total_vm_cycles = self.total_vm_cycles();
         let total_noops = self.total_noops();
@@ -150,7 +149,7 @@ pub fn analyze<A>(
     program: &str,
     stack_inputs: StackInputs,
     advice_provider: A,
-) -> Result<ProgramInfo, ProgramError>
+) -> Result<ExecutionDetails, ProgramError>
 where
     A: AdviceProvider,
 {
@@ -161,20 +160,20 @@ where
         .compile(program)
         .map_err(ProgramError::AssemblyError)?;
     let vm_state_iterator = processor::execute_iter(&program, stack_inputs, advice_provider);
-    let mut program_info = ProgramInfo::default();
+    let mut execution_details = ExecutionDetails::default();
 
     for state in vm_state_iterator {
         let vm_state = state.map_err(ProgramError::ExecutionError)?;
         if matches!(vm_state.op, Some(Operation::Noop)) {
-            program_info.incr_noop_count();
+            execution_details.incr_noop_count();
         }
         if let Some(asmop_info) = vm_state.asmop {
-            program_info.record_asmop(asmop_info);
+            execution_details.record_asmop(asmop_info);
         }
-        program_info.set_total_vm_cycles(vm_state.clk);
+        execution_details.set_total_vm_cycles(vm_state.clk);
     }
 
-    Ok(program_info)
+    Ok(execution_details)
 }
 
 // ASMOP STATS
@@ -233,7 +232,7 @@ impl AsmOpStats {
 
 #[cfg(test)]
 mod tests {
-    use super::{AsmOpStats, ProgramInfo, StackInputs};
+    use super::{AsmOpStats, ExecutionDetails, StackInputs};
     use miden::MemAdviceProvider;
 
     #[test]
@@ -242,9 +241,9 @@ mod tests {
             "proc.foo.1 loc_store.0 end begin mem_storew.1 dropw push.17 push.1 movdn.2 exec.foo end";
         let stack_inputs = StackInputs::default();
         let advice_provider = MemAdviceProvider::default();
-        let program_info = super::analyze(source, stack_inputs, advice_provider)
+        let execution_details = super::analyze(source, stack_inputs, advice_provider)
             .expect("analyze_test: Unexpected Error");
-        let expected_program_info = ProgramInfo {
+        let expected_details = ExecutionDetails {
             total_vm_cycles: 23,
             total_noops: 2,
             asm_op_stats: vec![
@@ -255,7 +254,7 @@ mod tests {
                 AsmOpStats::new("push".to_string(), 2, 3),
             ],
         };
-        assert_eq!(program_info, expected_program_info);
+        assert_eq!(execution_details, expected_details);
     }
 
     #[test]
@@ -264,9 +263,9 @@ mod tests {
         let stack_inputs = vec![1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
         let stack_inputs = StackInputs::try_from_values(stack_inputs).unwrap();
         let advice_provider = MemAdviceProvider::default();
-        let program_info = super::analyze(source, stack_inputs, advice_provider);
+        let execution_details = super::analyze(source, stack_inputs, advice_provider);
         let expected_error = "Execution Error: DivideByZero(1)";
-        assert_eq!(program_info.err().unwrap().to_string(), expected_error);
+        assert_eq!(execution_details.err().unwrap().to_string(), expected_error);
     }
 
     #[test]
@@ -274,8 +273,8 @@ mod tests {
         let source = "proc.foo.1 loc_store.0 end mem_storew.1 dropw push.17 exec.foo end";
         let stack_inputs = StackInputs::default();
         let advice_provider = MemAdviceProvider::default();
-        let program_info = super::analyze(source, stack_inputs, advice_provider);
+        let execution_details = super::analyze(source, stack_inputs, advice_provider);
         let expected_error = "Assembly Error: ParsingError(\"unexpected token: expected 'begin' but was 'mem_storew.1'\")";
-        assert_eq!(program_info.err().unwrap().to_string(), expected_error);
+        assert_eq!(execution_details.err().unwrap().to_string(), expected_error);
     }
 }